+++ lib/nostr_todo_sdk.dart
@@ -2,4 +2,5 @@
 library;
 
 export 'src/models/todo.dart';
+export 'src/nostr_todo_constants.dart';
 export 'src/services/todo_service.dart';